fixing page file and route onboarding / login
